More cleanup and fix expected response
